# swagger.yaml
---
openapi: "3.0.0"
info:
  title: Job Controller
  version: "V1.0.0"
  description: >-
    Run, monitor, and view logs from long running processes on remote hosts
  contact:
    name: John Stile
    email: john@stilen.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://127.0.0.1/V1
    description: Development server
  - url: https://staging.gigantic-server.com
    description: Staging server
  - url: https://api.gigantic-server.com
    description: Production server
tags:
  - name: auth
    description: >-
      Operations about authentication
  - name: stations
    description: >-
      Station where jobs will run
  - name: configs
    description: >-
      Operations about configs
  - name: jobs
    description: >-
      Operations about jobs
  - name: logs
    description: >-
      Operations about logs
  - name: users
    description: >-
      Operations about user
paths:
  "/login":
    post:
      tags:
        - auth
      summary: Authenticate user
      description: >-
        On successful Login,
        <br>1) copy the "Response body",
        <br>2) Click "Authorize" in upper right of this interface,
        <br>3) Paste the copied body into the form that pops up
        <br>4) Now protected routes will work (the lock will show unlocked)
      operationId: postLogin
      requestBody:
        description: Authentication credentials
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/PostAuth"
      security:
        - api_key: []
      responses:
        "202":
          $ref: "#/components/auth/Created"
        "401":
          $ref: "#/components/auth/BadRequestError"
        "403":
          $ref: "#/components/auth/NotFound"
        "503":
          $ref: "#/components/auth/ServiceUnavailable"
  "/auth_check":
    get:
      summary: Get current credentials
      description: >-
        Requires a valid JWT for HTTP 200.
        Returns authenticated user and claims as json
      operationId: getStations
      tags:
        - auth
      responses:
        "200":
          description: An Object with credentials
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/AuthCheck"
        "401":
          $ref: "#/components/responses/AuthorizationError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
      security:
        - {}
        - bearerAuth: []
  "/stations":
    get:
      summary: Get a list of stored stations
      description: >-
        Returns a list of station type objects, station type objects map
        a MAC address to alias and other info about where the job will run
      operationId: getStations
      tags:
        - stations
      responses:
        "200":
          description: An array of station objects
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/ArrayOfStations"
        "401":
          $ref: "#/components/responses/AuthorizationError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
      security:
        - api_key: []
    post:
      tags:
        - stations
      summary: Add a station
      operationId: addStation
      requestBody:
        description: Station, without unique id produced by server
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/PostStations"
      security:
        - api_key: []
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/ConflictError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  "/stations/{stationId}":
    put:
      tags:
        - stations
      summary: Change a station
      operationId: updateStation
      parameters:
        - "$ref": "#/components/parameters/stationIdParam"
      requestBody:
        description: Station, without unique id produced by server
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/PostStations"
      responses:
        "202":
          $ref: "#/components/responses/Accepted"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/ConflictError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
      security:
        - api_key: []
    delete:
      tags:
        - stations
      summary: Deletes a station
      operationId: deleteStations
      produces:
        - application/json
      parameters:
        - "$ref": "#/components/parameters/stationIdParam"
      responses:
        "202":
          $ref: "#/components/responses/Accepted"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/ConflictError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
      security:
        - api_key: []
components:
  parameters:
    stationIdParam:
      name: stationId
      in: path
      description: User set Station Identifier
      required: false
      type: string
  schemas:
    PostAuth:
      type: object
      properties:
        username:
          description: User Name
          type: string
        password:
          description: User Password
          type: string
      example: {"username": "john", "password": "password"}
    AuthCheck:
      type: object
      properties:
        user:
          description: User Name
          type: string
        claims:
          description: User Permissions
          type: string
    PostStations:
      type: object
      properties:
        ManufacturingSite:
          description: Physical location of station
          type: string
        JobType:
          description: Type of jobs run on this station
          type: string
        StationID:
          description: Family of jobs run on this station
          type: string
        NetToSerialMac:
          description: MAC address of station Nic
          type: string
    ArrayOfStations:
      type: array
      items:
        type: object
        properties:
          ManufacturingSite:
            type: string
          JobType:
            type: string
          StationID:
            type: string
          NetToSerialMac:
            type: string
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        userStatus:
          type: integer
          format: int32
          description: User Status
      xml:
        name: User
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
  responses:
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
    Updated:
      description: Resource updated
    Deleted:
      description: Resource deleted
    AuthorizationError:
      description: Access token is missing, expired, or invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            example: { "message": "Token has expired" }
    Accepted:
      description: Request accepted
      content:
        application/json:
          schema:
            "$ref": "#/components/responses/SuccessResponse"
          example: { "code": 202, "message": "Success" }
    Created:
      description: Resoure created
      content:
        application/json:
          schema:
            "$ref": "#/components/responses/SuccessResponse"
          example: { "code": 201, "message": "Success" }
    NotFound:
      description: No matching resource found for the given criteria.
    BadRequestError:
      description: Bad Request. Invalid JSON or keys.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { "code": 400, "message": "Bad Request." }
    ConflictError:
      description: Request conflict with current state of the server
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { "code": 409, "message": "Bad Request." }
    ServiceUnavailable:
      description: DataService not accessible
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { "code": 503, "message": "DataService not accessible" }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

...